decompile & match SailSea, SailCamera & SailPauseMenu
other things:
- rename some sprite-related enums & variables
- rename some task flags to better match their use cases
- add enum for task pause levels for readability purposes
- start work on more SailCommonObjects functions
- rename some parameters in obj.c/h to be clearer

+++ include/sail/sailVoyageManager.h
@@ -141,7 +141,7 @@ u8 SailVoyageManager__SeaMapObjectUnknownType3(SeaMapVoyagePathConfigNodeLink *w
 void SailVoyageManager__SetupVoyage(void);
 s32 SailVoyageManager__GetVoyagePos(void);
 VecFx32 *SailVoyageManager__GetVoyageVelocity(void);
-s32 SailVoyageManager__GetVoyageAngle(void);
+u16 SailVoyageManager__GetVoyageAngle(void);
 s32 SailVoyageManager__GetSegmentSize(SailVoyageSegment *segment);
 void SailVoyageManager__Main(void);
 void SailVoyageManager__Func_2157C34(SailVoyageManager *work);
